+++ lib/runtime/client/dev.ts
@@ -12,7 +12,18 @@ export function cssHotReload() {
             const originalSuccess = reporter.success
 
             reporter.success = (...args: any[]) => {
-                const linkNode = document.getElementById('css-main') as HTMLLinkElement
+                const linkNodes = document.getElementsByTagName('link')
+                let linkNode: HTMLLinkElement | undefined
+
+                // can't use for-of because it is not an array
+                // tslint:disable-next-line prefer-for-of
+                for (let i = 0; i < linkNodes.length; i++) {
+                    const node = linkNodes[i]
+                    if (node.href.indexOf('/main') !== -1) {
+                        linkNode = node
+                        break
+                    }
+                }
 
                 if (!linkNode) {
                     console.error('Error in CSS hot reload: link element not found!')
@@ -25,7 +36,7 @@ export function cssHotReload() {
                     parentNode.appendChild(newLink)
 
                     // defer removing the old style to prevent flash of unstyled content
-                    setTimeout(() => parentNode.removeChild(linkNode), 1000)
+                    setTimeout(() => parentNode.removeChild(linkNode as HTMLLinkElement), 1000)
                 }
 
                 originalSuccess.apply(reporter, args)