+++ test/bin/start.test.ts
@@ -2,6 +2,7 @@ import clean from '../../lib/bin/clean'
 import build from '../../lib/bin/build'
 import start from '../../lib/bin/start'
 import { delay } from '../../lib/runtime/util/time'
+import { waitForConnection } from '../../lib/runtime/util/network'
 import { getTestPort } from '../test-helpers'
 
 // Increase test timeout because builds might take a while
@@ -9,13 +10,25 @@ import { getTestPort } from '../test-helpers'
 
 describe('bin/start', () => {
 
-    it('will start', async () => {
+    beforeAll(async () => {
         await clean()
         await build()
+    })
+
+    it('will start the server', async () => {
         const port = await getTestPort()
         process.env.PORT = port
         const childProcess = await start('watch', 'fast', 'prod', 'debug')
-        await delay()
+        await waitForConnection(port)
+        childProcess.kill()
+    })
+
+    // can't test in TypeScript land because it requires the internal server in JavaScript
+    it.skip('will start the client', async () => {
+        const port = await getTestPort()
+        process.env.PORT = port
+        const childProcess = await start('prod', 'client')
+        await waitForConnection(port)
         childProcess.kill()
     })